Add charset option
The library shouldn't have assumed that UTF-8 was always the desired
character set. The charset option can now be set and we only attempt
to convert the charset if it doesn't match what's set as the option.

The option defaults to utf-8, and any conversion can be disabled by
setting it to false