Work around SplFileObject::DROP_NEW_LINE bug that strips first new line from each field
If a field contains new lines, the DROP_NEW_LINE flag strips the first new line from
every field in the file. To workaround this I have removed the DROP_NEW_LINE flag from
the setFlags call and added a check to see if the current row is an array with a single
value of null.

See https://bugs.php.net/bug.php?id=69181